# api/openapi-specs/openapi.yaml
openapi: 3.0.3
info:
  title: Users API
  description: Users basic CRUD (create, read, update and delete)
  version: v1.0.0
  x-service: users-service
  contact:
    email: rcebrian@github.com
  license:
    name: MIT
    url: 'https://www.mit.edu/~amini/LICENSE.md'
servers:
  - url: 'http://localhost:8080'
    description: localhost environment
tags:
  - name: system
    description: System health operations
  - name: users
    description: Users use cases
paths:
  /users:
    get:
      tags:
        - users
      operationId: getAllUsers
      summary: Get all users
      description: Retrieve all users from data storage
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDto'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    post:
      summary: Create user
      operationId: createUser
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: string
        '500':
          description: Internal Server Error
      description: Save user into data storage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
      tags:
        - users
  '/users/{userId}':
    get:
      summary: Get user by id
      tags:
        - users
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      operationId: getUserById
      description: Retrieve user information from ID
    parameters:
      - schema:
          type: string
        name: userId
        in: path
        required: true
        description: Numeric ID of the user to get
components:
  schemas:
    UserDto:
      title: UserDto
      type: object
      description: 'User model for API operations'
      properties:
        id:
          type: string
        name:
          type: string
        firstname:
          type: string